Fixed a bug in NSA LaRC cloud product preprocess:
We use the observation time to decide which satellite the data come from, for example:

!  GOES EAST  : 1815, 1845, 1915
!  GOES WEST  : 1830, 1900, 2030

But found that the GOES EAST time was changed to the same as GOES WEST.
This cause some preprocess can only find one of the satellite.
To fix the problem, the satellite id is added to help identify the data resources.
With this fix, both EAST and WEST should be able to get used in the preprocess.